move validation of key length from certificate parsing into validation of cert
